feat: SmoothMoving +Couroutine +DebuggingButton

+++ app/src/main/java/com/unnamed/mobile/component/model/Component.kt
@@ -6,9 +6,9 @@ open class Component{
     open val src: Int = 0
 }
 
-open class Dynamic(location: Pair<Int, Int>): Component(){
-    var location: Pair<Int, Int> = location
-    fun move(next: Pair<Int, Int>){
+open class Dynamic(location: Pair<Float, Float>): Component(){
+    var location: Pair<Float, Float> = location
+    fun move(next: Pair<Float, Float>){
         location = next
     }
 }
@@ -17,10 +17,10 @@ open class Static(location: Pair<Int, Int>): Component(){
     val location: Pair<Int, Int> = location
 }
 
-class Robot(location: Pair<Int, Int>): Dynamic(location){
+class Robot(location: Pair<Float, Float>): Dynamic(location){
     override val src = R.drawable.robot
-    var trace: MutableList<Pair<Int, Int>> = mutableListOf()
-    fun moveTo(next: Pair<Int, Int>){
+    var trace: MutableList<Pair<Float, Float>> = mutableListOf()
+    fun moveTo(next: Pair<Float, Float>){
         trace.add(next)
         move(next)
     }
